Refactor namespaces for clarity and maintainability
- Replace `TaskFlow.Api.Configuration` with `TaskFlow.Api.Extensions`
  to better group extension methods under a relevant namespace.
- Update test namespaces to align with the main code changes.
- Move provider-related functionality to `TaskFlow.Api.Providers`
  for improved separation of concerns.
- Simplify type names in `SwaggerDefaultValues` for better readability.
- Update `Program.cs` to use the new `Extensions` namespace.
- Adjust `using` directives across files to reflect namespace changes.
- Enhance project organization to improve maintainability and clarity.
