+++ TaskFlow.Api/Extensions/ApiVersioningServiceExtensions.cs
@@ -1,6 +1,6 @@
 using Asp.Versioning;
 
-namespace TaskFlow.Api.Configuration;
+namespace TaskFlow.Api.Extensions;
 
 /// <summary>
 /// Extension methods for configuring API versioning services
